Resolving legacy TODOs
The project format version is now v0.17.

The output directory suffix of a module is now deprecated and the name of
the module is used instead.

The user is now warned that the project format will be updated if it is
saved. In previous versions the format was updated, with the user's
permission when the project was loaded.

The access specifier of a super-class is always generated in a .sip file.
In previous versions 'public' was omitted to be compatible with old
versions of sip.

Resolves #6