Platform-specific header directory properties
The header directory properties that specify the header files to scan (ie.
the suffix directory and the file filter) have now been combined into a
single header file pattern.

The properties of a header directory (ie. the header file pattern and the
C++ parser arguments) can now be specified in the scanner tool GUI on a
per-platform basis.

Resolves #7